+++ serial-version/m_step.h
@@ -2,6 +2,6 @@
 #define EM_PROJECT_M_STEP_SERIAL
 
 // run maximization step
-void m_step(float *X, float *mean, float *cov, float *weights, float *p_val, int K, int N, int D);
+void m_step(double *X, double *mean, double *cov, double *weights, double *p_val, int K, int N, int D);
 
 #endif
